requirements.txt
fastapi
uvicorn
python-multipart
python-jose
passlib[bcrypt]
pydantic
pydantic-settings
mysql-connector-python
python-dotenv
google-generativeai==0.3.2
docxtpl
python-docx
Pillow